trying to get my refactoring to work
