+++ .gitignore
@@ -1,7 +1,6 @@
 /venv
-/examples/Heisenberg_Chain/cache
-/tests/__pycache__
-/examples/Heisenberg_Chain/__pycache__
+/Heisenberg_Chain/cache
+__pycache__/
 /echostate/__pycache__
 /logs
-/true_hist
+/echostate.egg-info
