+++ echostate/__init__.py
@@ -9,13 +9,12 @@ Public API:
 from .core.esn import ESN
 from .core.reservoir import Reservoir
 from .core.trainer import Trainer
-from .core.metrics import mean_absolute_error, mean_squared_error
-from .io.serialization import save_esn, load_esn
+from .utils import mean_absolute_error, mean_squared_error
+
 
 __all__ = [
     "ESN", "Reservoir", "Trainer",
     "mean_absolute_error", "mean_squared_error",
-    "save_esn", "load_esn",
 ]
 
 __version__ = "0.1.0"
