+++ echostate/core/__init__.py
@@ -1,6 +1,6 @@
 from .esn import ESN
 from .reservoir import Reservoir
 from .trainer import Trainer
-from .metrics import mean_absolute_error, mean_squared_error
 
-__all__ = ["ESN", "Reservoir", "Trainer", "mean_absolute_error", "mean_squared_error"]
+
+__all__ = ["ESN", "Reservoir", "Trainer"]
