+++ echostate/core/esn.py
@@ -5,8 +5,8 @@ import torch
 
 from .reservoir import Reservoir
 from .trainer import Trainer
-from .metrics import mean_absolute_error, mean_squared_error
-from ..logging.utils import log_tensor, tensor_stats
+from ..utils import mean_absolute_error, mean_squared_error
+from ..esn_logging.utils import log_tensor, tensor_stats
 
 LOGGER = logging.getLogger(__name__)
