--- echostate/echostate.egg-info/SOURCES.txt
@@ -1,23 +1,0 @@
-README.md
-setup.py
-core/__init__.py
-core/esn.py
-core/metrics.py
-core/reservoir.py
-core/trainer.py
-echostate.egg-info/PKG-INFO
-echostate.egg-info/SOURCES.txt
-echostate.egg-info/dependency_links.txt
-echostate.egg-info/requires.txt
-echostate.egg-info/top_level.txt
-logging/__init__.py
-logging/config.py
-logging/utils.py
-tests/__init__.py
-tests/test_esn.py
-tests/test_reservoir.py
-tests/test_trainer.py
-tests/test_utils.py
-tuning/__init__.py
-tuning/optuna_tuner.py
-tuning/search_spaces.py
