--- echostate/echostate.egg-info/requires.txt
@@ -1,10 +1,0 @@
-matplotlib
-numpy
-optuna
-pandas
-pytest
-qutip
-safetensors
-scipy
-setuptools
-torch
