--- echostate/echostate.egg-info/top_level.txt
@@ -1,4 +1,0 @@
-core
-logging
-tests
-tuning
