+++ pyproject.toml
@@ -19,3 +19,4 @@ dependencies = [
 
 [tool.setuptools]
 packages = ["echostate"]
+
